fix: POST /api/v1/configs API認証エラーとTypeScript型安全性の修正
- 認証チェックのキー名を app.config_password に統一
- POST API テストの認証設定を共通化
- any/unknown型を適切な型定義に修正
  - JsonValue型の導入でJSON値の型安全性向上
  - Zodスキーマの型定義を具体化
  - テストファイルの型定義を修正
- 全テスト(53/53)がパス、TypeScriptエラーゼロ達成
