+++ src/api/index.ts
@@ -15,6 +15,7 @@ import storagesDeleteHandler from './v1/storages/delete';
 import authRouter, { callbackRoute, logoutRoute, meRoute } from './v1/auth';
 import { callbackHandler } from './v1/auth/callback/get';
 import { logoutHandler } from './v1/auth/logout/post';
+import configsRouter from './v1/configs/route';
 import { meHandler } from './v1/auth/me/get';
 import type { Env } from '@/types/env';
 
@@ -66,7 +67,6 @@ v1.route('/storages', storagesRouter);
 v1.route('/auth', authRouter);
 
 // Config routes
-import configsRouter from './v1/configs/handlers';
 v1.route('/configs', configsRouter);
 
 // Mount v1 API routes
