+++ test/api/v1/configs/get.test.ts
@@ -15,7 +15,7 @@ interface ConfigItem {
   type: 'string' | 'boolean' | 'number' | 'json';
   description: string | null;
   system_config: boolean;
-  validation: any;
+  validation: Record<string, string | number | boolean> | null;
   created_at: string | null;
   updated_at: string | null;
 }
@@ -76,7 +76,7 @@ describe('GET /api/v1/configs/:key', () => {
         type: 'string',
         description: 'Google OAuth2 Client ID',
         system_config: expect.any(Boolean),
-        validation: expect.toSatisfy((val: any) => val === null || typeof val === 'object'),
+        validation: expect.toSatisfy((val: Record<string, string | number | boolean> | null) => val === null || typeof val === 'object'),
         created_at: expect.any(String),
         updated_at: expect.any(String)
       });
